Backend/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
build==1.2.2.post1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
envparse==0.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pip-tools==7.4.1
psycopg2==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
pyproject_hooks==1.2.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
sentry-sdk==2.22.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
